Make mobile app functional and remove tab nav

+++ mobile/navigation/tabs.js
@@ -1,16 +1,16 @@
 import React, { useState, useCallback } from 'react';
-import { StyleSheet, Text, View, Image, TouchableOpacity } from 'react-native';
+import { StyleSheet, Text, View, Image, TouchableOpacity, Button, Alert } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-import HomeScreen from '../screens/HomeScreen.js';
+import Login from '../screens/Login.js';
 import RegistrationScreen from '../screens/RegistrationScreen.js';
 import TodoScreen from '../screens/TodoScreen.js';
 import DeskScreen from '../screens/DeskScreen.js';
 import Landing from '../screens/Landing.js';
 
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { Button } from 'react-native-elements/dist/buttons/Button';
+import HomeScreen from '../screens/Login.js';
 
 
 const RegistrationStack = createNativeStackNavigator();
@@ -20,59 +20,66 @@ function RegistrationStackScreen() {
 		<RegistrationStack.Navigator
 			initialRouteName="Login"
 			screenOptions={{ headerShown: true }}>
-			<RegistrationStack.Screen name="Login" component={HomeScreen} />
+			<RegistrationStack.Screen name="Login" component={Login} />
 			<RegistrationStack.Screen name="Registration" component={RegistrationScreen} />
 		</RegistrationStack.Navigator>
 	);
 }
 
-
-const MainStack = createNativeStackNavigator();
-
-function MainStackScreen() {
-	return (
-		<MainStack.Navigator
-			initialRouteName="Todo"
-			screenOptions={{ headerShown: true }}>
-			<MainStack.Screen name="Todo" component={Landing} />
-		</MainStack.Navigator>
-	);
-}
-
 const DeskStack = createNativeStackNavigator();
 
-function DeskStackScreen() {
+	
+function userLogout(navigation) {
+	Alert.alert(
+		"Flourish Logout",
+		"Are you sure you want to logout?",
+		[
+		  {
+			text: "Logout",
+			onPress: () => {
+					 AsyncStorage.setItem("token", "")
+					.then(() => {
+						AsyncStorage.setItem("user_data", "")
+						.then(() => {
+							navigation.navigate("Login")
+						}).catch(e=>alert(e.message));
+					}).catch(e=>alert(e.message));
+			},
+			style: "destructive",
+		  },
+		  {
+			text: "Cancel",
+			style: "Neutral",
+		  },
+		  
+		]
+	  );
+}
+
+function DeskStackScreen({navigation}) {
 	return (
 		<DeskStack.Navigator
 			initialRouteName="Home"
 			screenOptions={{ headerShown: true }}>
-			<DeskStack.Screen name="Lists" component={DeskScreen} />
+			<DeskStack.Screen name="Lists" component={DeskScreen} options={{
+				headerRight: () => (
+				  <Button
+					onPress={() => userLogout(navigation)}
+					title="Logout"
+				  />
+				),
+			  }} />
 			<DeskStack.Screen name="Todo" component={TodoScreen} />
 		</DeskStack.Navigator>
 	);
 }
 
-const Tab = createBottomTabNavigator();
-
+const Stack = createNativeStackNavigator();
 
 function Tabs(props) {
 
-	// const onRefresh = useCallback(() => {
-	// 	props.setUser(true);
-	// 	wait(2000).then(() => props.setUser(false));
-	//   }, []);
-	function logoutUser() {
-		AsyncStorage.setItem("token", "")
-		.then(() => {
-			AsyncStorage.setItem("user_data", "")
-				.then(() => {
-					props.setUser(true);
-				});
-		});
-	}
-
-	return true ? (
-		<Tab.Navigator
+	return (
+		<Stack.Navigator
 			screenOptions={{
 				"headerShown": false,
 				"tabBarShowLabel": false,
@@ -82,62 +89,11 @@ function Tabs(props) {
 					},
 					null
 				]
-			}}>
-			<Tab.Screen name="Home" component={MainStackScreen} options={{
-				tabBarIcon: ({ focused }) => (
-					<View style={{ alignItems: 'center', justifyContent: 'center', top: 10 }}>
-						<Image
-							source={require('../assets/icons/home.png')}
-							resizeMode="contain"
-							style={{
-								width: 25,
-								height: 25,
-								tintColor: focused ? '#64706f' : '#8BA2A1'
-							}} />
-						<Text style={{ color: focused ? '#64706f' : '#8BA2A1', fontSize: 12 }}
-						>Home</Text>
-					</View>
-				),
-			}} />
-			<Tab.Screen name="Desk" component={DeskStackScreen} options={{
-				tabBarIcon: ({ focused }) => (
-					<View style={{ alignItems: 'center', justifyContent: 'center', top: 10 }}>
-						<Image
-							source={require('../assets/icons/work.png')}
-							resizeMode="contain"
-							style={{
-								width: 25,
-								height: 25,
-								tintColor: focused ? '#64706f' : '#8BA2A1'
-							}} />
-						<Text style={{ color: focused ? '#64706f' : '#8BA2A1', fontSize: 12 }}
-						>Desk</Text>
-					</View>
-				),
-			}} /><Tab.Screen name="Register" component={RegistrationStackScreen} options={{
-				tabBarIcon: ({ focused }) => (
-					<View onPress={() => { setUser(true) }} style={{ alignItems: 'center', justifyContent: 'center', top: 10 }}>
-						<Image
-							source={require('../assets/icons/user.png')}
-							resizeMode="contain"
-							style={{
-								width: 25,
-								height: 25,
-								tintColor: focused ? '#64706f' : '#8BA2A1'
-							}} />
-						<Text style={{ color: focused ? '#64706f' : '#8BA2A1', fontSize: 12 }}
-						>Register</Text>
-					</View>
-				),
-			}} />
-		</Tab.Navigator>
-	) : <>
-		<Button
-			onPress={logoutUser}
-			title="Logout"
-			color="dd0000"
-		/>
-	</>;
+			}} 
+			>
+			<Stack.Screen name="Register" component={RegistrationStackScreen} />
+			<Stack.Screen name="Desk" component={DeskStackScreen} />
+		</Stack.Navigator>
+	);
 }
-
 export default Tabs;
